+++ src/server/axum/layers/prometheus.rs
@@ -4,10 +4,14 @@ use axum::body::Body;
 use axum::extract::MatchedPath;
 use axum::http::Request;
 use axum::response::Response;
+use bytesize::ByteSize;
 use futures::future::BoxFuture;
-use metrics::{counter, histogram};
+use metrics::{counter, gauge, histogram};
+use std::fmt;
+use std::path::Path;
 use std::task::{Context, Poll};
 use std::time::Instant;
+use sysinfo::{Disks, System};
 use tower::{Layer, Service};
 
 /// Prometheus metrics layer for Axum
@@ -49,7 +53,6 @@ where
     }
 
     fn call(&mut self, request: Request<Body>) -> Self::Future {
-        let start = Instant::now();
         let path = if let Some(matched_path) = request.extensions().get::<MatchedPath>() {
             matched_path.as_str().to_owned()
         } else {
@@ -58,6 +61,7 @@ where
         let method = request.method().to_string();
         let service_name = self.service_name.clone();
 
+        let start = Instant::now();
         let future = self.inner.call(request);
         Box::pin(async move {
             let response = future.await?;
@@ -69,15 +73,116 @@ where
                 let labels = [
                     ("method", method),
                     ("path", path),
-                    ("service", service_name),
+                    ("service", service_name.clone()),
                     ("status", status),
                 ];
 
                 counter!("http_requests_total", &labels).increment(1);
                 histogram!("http_requests_duration_seconds", &labels).record(latency);
+
+                // System metrics
+                let system_metrics = SystemMetrics::new("/").await;
+                system_metrics.add_metrics(service_name);
             }
 
             Ok(response)
         })
+    }
+}
+
+#[derive(Debug, Clone)]
+struct SystemMetrics {
+    /// Average CPU usage in percent
+    cpu_usage: f32,
+
+    /// Total memory in bytes
+    total_memory: u64,
+
+    /// Used memory in bytes
+    used_memory: u64,
+
+    /// Total swap space in bytes
+    total_swap: u64,
+
+    /// Used swap space in bytes
+    used_swap: u64,
+
+    /// Total disk space in bytes for a specified mount point
+    total_disks_space: u64,
+
+    /// Used disk space in bytes for a specified mount point
+    used_disks_space: u64,
+}
+
+impl SystemMetrics {
+    async fn new(disk_mount_point: &str) -> Self {
+        let mut sys = System::new_all();
+
+        // CPU
+        sys.refresh_cpu_usage();
+        let mut cpu_usage = sys.global_cpu_usage();
+        tokio::time::sleep(sysinfo::MINIMUM_CPU_UPDATE_INTERVAL).await;
+        sys.refresh_cpu_usage();
+        cpu_usage += sys.global_cpu_usage();
+        cpu_usage /= 2.0;
+
+        // Memory
+        sys.refresh_memory();
+        let total_memory = sys.total_memory();
+        let used_memory = sys.used_memory();
+
+        // Swap
+        let total_swap = sys.total_swap();
+        let used_swap = sys.used_swap();
+
+        // Disks
+        let disks = Disks::new_with_refreshed_list();
+        let mut total_disks_space = 0;
+        let mut used_disks_space = 0;
+        for disk in &disks {
+            if disk.mount_point() == Path::new(disk_mount_point) {
+                total_disks_space += disk.total_space();
+                used_disks_space += disk.total_space() - disk.available_space();
+            }
+        }
+
+        Self {
+            cpu_usage,
+            total_memory,
+            used_memory,
+            total_swap,
+            used_swap,
+            total_disks_space,
+            used_disks_space,
+        }
+    }
+
+    fn add_metrics(&self, service_name: String) {
+        gauge!("system_cpu_usage", "service" => service_name.clone()).set(self.cpu_usage);
+        gauge!("system_total_memory", "service" => service_name.clone()).set(self.total_memory as f64);
+        gauge!("system_used_memory", "service" => service_name.clone()).set(self.used_memory as f64);
+        gauge!("system_total_swap", "service" => service_name.clone()).set(self.total_swap as f64);
+        gauge!("system_used_swap", "service" => service_name.clone()).set(self.used_swap as f64);
+        gauge!("system_total_disks_space", "service" => service_name.clone()).set(self.total_disks_space as f64);
+        gauge!("system_used_disks_usage", "service" => service_name).set(self.used_disks_space as f64);
+    }
+}
+
+impl fmt::Display for SystemMetrics {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "CPUs:       {:.1}%\n\
+             Memory:     {} / {}\n\
+             Swap:       {} / {}\n\
+             Disk usage: {} / {}",
+            self.cpu_usage,
+            ByteSize::b(self.used_memory),
+            ByteSize::b(self.total_memory),
+            ByteSize::b(self.used_swap),
+            ByteSize::b(self.total_swap),
+            ByteSize::b(self.used_disks_space),
+            ByteSize::b(self.total_disks_space),
+        )
     }
 }